Break Eli into smaller parts
First off, File and Buffer were basically the same thing, but file wasn't actually tied to an
actual file on the disk, so those have been combined.

A lot of the code in eli.c has been moved to a context file named editor.c (not sure if that
name is appropriate yet).