Give Default Actions a next mode
The default action member has been removed from the Mode struct and moved to the bottom of the action list.
By putting the default action in the action list, we will treat it like any other action, which means it will have a next mode.

The goal here is to make it easier to create a command mode whose default action will probably
be NULL, but nextmode will be NORMAL, i.e. command mode will live for one action (default) and
go back to normal mode.