Buffer: Take a name on new, read, and write functions
Buffers are pretty tightly coupled to their names, which made it so you had to set a buffer's
name before calling certain functions, which can be annoying.

This change makes the buffer name a more integral part of the buffer object.
You now must give a buffer name or NULL when creating a buffer.
You must give a name or NULL when reading or writing a buffer, this name can be different from the buffer's name, and if no name is given the buffer's name is used.
Clearing the buffer before reading has also been moved into the buf_read function, since these two functions almost needed to be called in this order, i.e. clear before read.